Reorganisation of urls, sites, admin.py
