+++ observations/templates/admin/change_list.html
@@ -51,7 +51,7 @@
   <span class="pull-right">
     {% block object-tools-items %}
       {% url cl.opts|admin_urlname:'add' as add_url %}
-      <a href="{% add_preserved_filters add_url is_popup %}" class="btn btn-primary">
+      <a href="{% add_preserved_filters add_url is_popup to_field %}" class="btn btn-primary">
       <span class="glyphicon glyphicon-plus"></span> {% blocktrans with cl.opts.verbose_name as name %}Add {{ name }}{% endblocktrans %}
       </a>
     {% endblock %}
