+++ observations/templates/admin/index.html
@@ -98,7 +98,7 @@
   {% if perms.observations.change_penguinobservation %}<li><a href="{% url "admin:observations_penguinobservation_changelist" %}"><span class="glyphicon glyphicon-eye-open"></span> Observations</a></li>{% endif %}
   {% if perms.observations.change_penguincount %}<li><a href="{% url "admin:observations_penguincount_changelist" %}"><span class="glyphicon glyphicon-check"></span> Penguin count</a></li>{% endif %}
   {% if perms.auth.change_user %}<li><a href="{% url "admin:observations_penguinuser_changelist" %}"><span class="glyphicon glyphicon-user"></span> Users</a></li>{% endif %}
-    <li><a href="{% url "admin:observations_camera_changelist" %}"><span class="glyphicon glyphicon-camera"></span> Cameras</a></li>
+  <li><a href="{% url "admin:observations_camera_changelist" %}"><span class="glyphicon glyphicon-camera"></span> Cameras</a></li>
   {% endif %}
   {% if perms.auth.change_user %}<li><a href="{% url "s3_view" %}"><span class="glyphicon glyphicon-user"></span> S3 View</a></li>{% endif %}
             <li><a href="/help/"><span class="glyphicon glypthicon-question-sign"></span>Help </a>
@@ -136,7 +136,7 @@
     <div class="col-xs-2">
       <form class="form form-inline" role="form"><div class="form-group">
         {{ gform }} <br />
-        <input type="submit" class="btn"> 
+        <input type="submit" class="btn">
       </div></form>
     </div>
     <div class="col-xs-10">
